Convert HighlightAndZoomLineChart to tsx
